+++ RELEASES.md
@@ -39,6 +39,7 @@ Upcoming
     - Add getIbisManager to Configuration
 - Make it possible to stop and start all adapters per configuration
 - Add "*ALL*" link in "Show configuration status"
+- Add attribute transformNullMessage to PipeLine